requirements.txt
torch~=1.13.1
torchvision~=0.14.1
pylint~=2.16.2
Pillow~=9.4.0
albumentations~=1.3.0
numpy~=1.24.2
opencv-python~=4.7.0.72